bflibrary: Rename uppercase types to not clash with Winapi
We could instead just use definitions from Winapi. But we have
a history of using lower case type names, from KeeperFX.

+++ bflibrary/include/bffile.h
@@ -48,7 +48,7 @@ enum TbFileSeekMode { // type=int8_t
 
 typedef enum TbFileSeekMode TbFileSeekMode;
 
-typedef int32_t TbFileHandle;
+typedef ulong TbFileHandle;
 
 #if defined(_finddata_t) // GCC Windows API
 typedef struct _finddata_t TbFILE_FIND;
